Install middle click event handler on KXmlGui Toolbar menu
For ease of access, KXmlGui adds actions in a delayed/split toolbar
button to its own context menu. We have no access to this menu, so
we cannot install our middle-click event handler for the back/forward
action menus.

To address this, a horrible hack is used to install the event filter
on the associated widget when the action becomes visible.